+++ mrblib/acli/cli.rb
@@ -49,6 +49,8 @@ Options:
 
   --sub-protocol        # Custom WebSocket subprotocol
 
+  --msgpack             # Provide this switch to use Msgpack encoded messages for communication
+
   --quit-after          # Automatically quit after an even occured.
                         # Possible values are:
                         #  - connected — quit right after successful connection ("welcome" message)
@@ -74,7 +76,8 @@ Options:
         "channel-params:",
         "headers:",
         "sub-protocol:",
-        "quit-after:"
+        "quit-after:",
+        "msgpack"
       ).yield_self do |opts|
         print_version if opts["v"]
         print_help if opts["h"]
@@ -90,13 +93,19 @@ Options:
           channel = "#{channel} #{opts["channel-params"].gsub(",", " ")}"
         end
 
+        if opts["msgpack"]
+          opts["sub-protocol"] = "actioncable-v1-msgpack" if opts["sub-protocol"].empty?
+          opts[:coder] = Coders::Msgpack
+        end
+
         {
           url: opts["u"] || opts["url"],
           channel: channel,
           quit_after: opts["quit-after"],
           headers: headers,
-          protocol: opts["sub-protocol"]
-        }.tap { |data| data.delete_if { _2.empty? } }
+          protocol: opts["sub-protocol"],
+          coder: opts[:coder]
+        }.tap { |data| data.delete_if { _2.nil? || (_2.is_a?(String) && _2.empty?) } }
       end
     end
   end